chore: move success tick into its' own component
